Fix highlighting bugs
First, I was using undefined instead of null when trying to figure out
if something indexed passed the end of an array. That didn't work.

Second, the plugin seems to process highlight events before edit events
when the user clicks the x key. Normally this wouldn't be a problem,
but if there are tabs on the line, things would get very messed up.