Text decorations at top (#358)
* Add option to draw text decorations on top

Plugins like EasyMotion can be drawn with text decorations. This has a
lot of benefits, but it cases text to shift all over the place. This
commit adds an option to draw text decorations on top of their
associated text in a way that is much more in line with how
NeoVim/EasyMotion actually work.

* Small comment change

Co-authored-by: Jack Garner <jacksonhgarner@outlook.com>