Fix i_Ctrl-a
Problem:
It was isolating the text to be copied by trimming first from
the front of it, causing it to shift the position of that desired
text, and thus copying the incorrect text.

Fix:
By trimming first from the end of the inserted text (and secondly
from the start) the shift happens only after the trim, so the
text gets isolated correctly for it to be copied.