Fix normal mode freezing when switching windows
This fixes a race condition where, if the set of visible editors changes
while we're in syncLayout, syncLayout will still resolve
changeLayoutPromise, meaning that syncActiveEditor will attempt to try
setting the current window in Neovim. That will fail, since the Neovim
window was never set up in syncLayout, which leads to everything Neovim
related not working for that editor until you switch editors to force
syncActiveEditor to be rerun.

Fix that bug by avoiding resolving changeLayoutPromise if the set of
visible editors has changed, which avoids syncActiveEditor continuing
until syncLayout is rerun with the correct visible editors.

Also turn the failure to look up the Neovim window in syncActiveEditor
from a warning into an error, since it's a fairly clear sign that
something's wrong and it should never be reached under normal
circumstances.