fix(highlight): remove flash.nvim lag (#1389)
* fix(highlight): remove flash.nvim lag
The bug was caused due to the new behavior in 0.4.0 that can now render blank highlights, used for eol visual cursor and blank leap.nvim labels. However, flash.nvim adds its grey overlay to every single cell in the buffer, and so thousands of "blank" cells were trying to be rendered using text decoration. This commit makes it so it will only render highlights that are over text.

* fix(highlight): remove duplicate highlights and simplify logic