feat(buffer): add external buffer doc provider
This opens a document with a proper buffer name and matches it to the
appropriate neovim external buffer.

The document is "readonly" as far as VSCode is concerned, but it can
still be edited by setting `modifiable`, so we still need to sync
buffer changes whenever it updates. This can break highlighting, however.