refactor: move highlight grid into its own data structure
This is the first step of many, but I'm beginning to separate the logic
that stores the highlight positions themselves and the logic that sends
them to VSCode. This will facilitate easier unit testing of the actual
highlight logic.

Ideally, I would love to test the HighlightProvider as a whole,b ut we
can't test VSCode specific functions within our mocha tests