Update README.md for Snap users
I've installed nvim as a snap package and VSCode could not detect automatically. This is because the Ubuntu 2204 package is 0.6, which is further behind than this extension recommends. 

A bit of googling found that snap stores the binary at: /snap/nvim/current/usr/bin/nvim

Updated the README as people may install from Ubuntu Store which uses snap.