fix(buffers): combine buf+wincleanup into one call
This ensures we always do buffer cleanup after window cleanup, and also
makes it sort of "atomic", but we can still fire + forget the request in
case it takes too long.

Also add a simple log for the case where it takes longer than expected,
just so we can have some record of this scenario instead of silently
ignoring it.